move depsync into platform
