+++ cmd/mdmctl/setup.go
@@ -4,11 +4,11 @@ import (
 	"github.com/go-kit/kit/log"
 	httptransport "github.com/go-kit/kit/transport/http"
 
-	"github.com/micromdm/micromdm/dep/depsync"
 	"github.com/micromdm/micromdm/platform/appstore"
 	"github.com/micromdm/micromdm/platform/blueprint"
 	"github.com/micromdm/micromdm/platform/config"
 	"github.com/micromdm/micromdm/platform/dep"
+	"github.com/micromdm/micromdm/platform/dep/sync"
 	"github.com/micromdm/micromdm/platform/device"
 	"github.com/micromdm/micromdm/platform/profile"
 	"github.com/micromdm/micromdm/platform/remove"
@@ -24,7 +24,7 @@ type remoteServices struct {
 	configsvc    config.Service
 	appsvc       appstore.Service
 	depsvc       dep.Service
-	depsyncsvc   depsync.Service
+	depsyncsvc   sync.Service
 }
 
 func setupClient(logger log.Logger) (*remoteServices, error) {
@@ -89,7 +89,7 @@ func setupClient(logger log.Logger) (*remoteServices, error) {
 		return nil, err
 	}
 
-	depsyncsvc, err := depsync.NewHTTPClient(
+	depsyncsvc, err := sync.NewHTTPClient(
 		cfg.ServerURL, cfg.APIToken, logger,
 		httptransport.SetClient(skipVerifyHTTPClient(cfg.SkipVerify)))
 	if err != nil {
