+++ platform/dep/sync/apply_autoassigner.go
@@ -1,4 +1,4 @@
-package depsync
+package sync
 
 import (
 	"context"
