+++ platform/dep/sync/client.go
@@ -1,4 +1,4 @@
-package depsync
+package sync
 
 import (
 	"net/url"
