+++ platform/dep/sync/get_autoassigners.go
@@ -1,4 +1,4 @@
-package depsync
+package sync
 
 import (
 	"context"
